etcdctl: add diagnosis offline analysis
This commit adds the offline analysis functionality to the diagnosis command. This feature allows users to inspect and analyze an etcd data directory when the cluster is not running, which is crucial for disaster recovery and debugging.

The code for parsing the backend database (`db`) file and its revisions is migrated from the ahrtr/etcd-diagnosis repository.

Resolves #20275

Signed-off-by: Yerasala Venkata, Seshachalam <seshachalam.yerasala.venkata@sap.com>